Add async file conversion example to ConvertAPI
This commit introduces an asynchronous file conversion feature to the ConvertAPI .NET example solution. The feature is encapsulated in a new project, AsyncConversion. The feature works by first submitting a file for conversion, then frequently checking if the conversion is done until the file is converted. This will prevent blocking of the execution flow of the program and enhance performance. The entry point of the project is envisioned to be AsyncConversion/Program.cs.