Update ConvertUrlToPdf project to .NET 6.0
The existing ConvertUrlToPdf project was refactored from .NET 4.7 to .NET 6.0. The AssemblyInfo.cs file was also deleted, considering .NET Core and onwards does not require separate AssemblyInfo files, as most of the assembly information can be configured in the new csproj format. The project GUID was updated due to the creation of a new project file under .NET 6.0. In addition, replaced WebClient with HttpClient in the file ConvertUrlToPdf/Program.cs for better performance and cleaner code. Finally, the solution's project references were updated to reflect these changes.