Add file storage feature in ConvertAPI usage
In this commit, we introduced an option for users to store the converted file to ConvertAPI's file storage. A new boolean constant 'StoreFile' has been added. If this constant is set to true, the file will be stored for maximum 3 hours in ConvertAPI's storage. We modified the FileConverter's SubmitFileForConversion method to accept an additional argument 'storeFile'. Accordingly, ConvertAPI's API request has been updated to include 'StoreFile' in query if 'storeFile' equals true. We also handled the API response to read JSON if the content type is 'application/json' and show the stored file's URL. Three new classes 'ConvertApiResponse', 'ConvertApiFiles' and 'AsyncResponse' have been added to deserialize JSON response.